tui: refactor
